+++ MusicManager/Song.h
@@ -36,9 +36,9 @@ public:
 
     void setPath(const std::string &path);
 
-    float getRate() const;
+    int getRate() const;
 
-    void setRate(float rate);
+    void setRate(int rate);
 
     const std::string &getSongName() const;
